Individual to typescript
